requirements.txt
pandas
requests
pyyaml
python-dotenv
prometheus-client
pydantic>=2
pydantic-settings>=2
jinja2
pytest



httpx
sqlalchemy
aiosqlite
greenlet
tqdm
fastapi
uvicorn
sse-starlette
